refactor(app): refactored controllers, models and views
in the model methods using select was refactored to use pluck this is
because using select returns Active Record object which is very large.
Associated controllers and views that uses this method in the model were
then updated

this refactor was done to ensure we do not have our query returning
Active Record object that will occupy memory.